Added LARGE_STACK feature that when disabled (using #define MSCOMP_WITHOUT_LARGE_STACK) will cause various large variables to be heap-based instead of stack-based. Now in use by LZNT1Dictionary (512-768 kb reduction in stack usage) and HuffmanEncoder::CreateCodesSlow (512 kb reduction in stack usage, but only in very rare cases).
Also fixed some code to not produce warnings in MSVC.

+++ src/xpress_huff_compress.cpp
@@ -148,7 +148,7 @@ static size_t xh_compress_lz77(const_bytes in, int32_t /* * */ in_len, const_byt
 	return out - out_orig;
 }
 WARNINGS_POP()
-static uint32_t xh_compress_no_matching(const_bytes in, int32_t in_len, bool is_end, bytes out, uint32_t symbol_counts[SYMBOLS])
+static size_t xh_compress_no_matching(const_bytes in, size_t in_len, bool is_end, bytes out, uint32_t symbol_counts[SYMBOLS])
 {
 	const const_bytes in_end = in + in_len, in_endx = in_end - 32;
 	const const_bytes out_orig = out;
